script-captura/requirements.txt
dotenv
mysql-connector-python
psutil
slack_sdk